Refactored Sass
The Sass I used was a mess. I’ve refactored the code and split it up
into sections relevant to each module/area of the site.

Variables and typeface declarations are now in source.scss, and I’ve
greatly expanded the number of partials and mixins available to the
theme.