Added new article template
The “archive article” is a third alternative option for index and
search result posts. While I use a lot of my own photographs in and
around the blog, I’ve never used the “featured image” feature of
WordPress.

I extract the first image of the post along with the first sentence,
and set it all into a box together.